Add transport and serializer for local Unix syslog
New SyslogLocalMessageSerializer and SyslogLocalSender classes deal with
sending messages to a local Unix syslog via the standard openlog(), syslog()
and closelog() system calls. Naturally, this only works on Unix-like systems
such as Linux, BSD, Mac OS X, and so on: a CommunicationsException will be
thrown if users attempt to use the SyslogLocalSender class on Windows.